+++ Changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add _Checkstyle_ baseline generation using `./gradlew generateCheckstyleBaseline` task
 - Add support for pre-AndroidX dependencies
 - Fix Robolectric support for `./gradlew projectCoverage` task.
+- Add Versioning Plugin
 
 ## [0.3.0] - 2020-01-07